backPaciente/requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
PyMySQL==1.1.1
sqlparse==0.5.3
tzdata==2025.2
django-cors-headers
requests
cryptography